Latest code of select column panel added
